Wednesday 29th June
Groove image now changed to have a white background and transparent groove, meaning that when the colour of the callout box is changed, the groove changes with it. Example, the purple callout box on forum headers.  However, we can't adopt this method on the frontpage slider because the callout box here sits on top of a background image and not a white background.  So I have set the slider callout box to a fixed purple and groove image.  For all other callouts , the colour can be changed and the groove will change automatically.
Dear Joe font tested on iPad and Android.
Carousel slider tested on ipad, touch gestures also work for swiping navigation on slider
Box shadow removed on all border for better consistency
The cyan brushstroke has been renamed to cyan2 so that it doesn't try load the old one.  So the new brushstroke will show. (the new brushstroke sent by Joanna is a JPG, the old one was a PNG, Moodle tried to load the png by default.  If this size works I suggest we save all br